Game loop and enemies

+++ app/src/main/java/com/phatphoophoo/pdtran/herotyper/objects/Bullet.kt
@@ -1,6 +1,6 @@
 package com.phatphoophoo.pdtran.herotyper.objects
 
-class Bullet(override var position: Pair<Int,Int>, override var velocity: Float) : GameObject {
+class Bullet(override var position: Pair<Float, Float>, override var velocity: Float) : GameObject {
     override val direction: Direction = Direction.UP
     override val height: Float = 10f
     override val width: Float = 10f
