+++ app/src/main/java/com/phatphoophoo/pdtran/herotyper/objects/Player.kt
@@ -1,6 +1,6 @@
 package com.phatphoophoo.pdtran.herotyper.objects
 
-class Player(override var position: Pair<Int, Int>) : GameObject {
+class Player(override var position: Pair<Float, Float>) : GameObject {
     override val direction = Direction.NONE
     override var velocity: Float = 0f
     override val height: Float = 40f
